+++ src/py_clean_cli/__init__.py
@@ -1,5 +1,6 @@
 from .decorators import command
 from .models import CommandArgsModel
+from .main import package_cli
 
-__version__ = "0.1.0"
-__all__ = ["command", "CommandArgsModel", "__version__"]
+__version__ = "0.1.1"
+__all__ = ["command", "CommandArgsModel", "__version__", "package_cli"]